requirements.txt
Django>=4.2.0
django-crispy-forms>=2.0
crispy-bootstrap4>=0.7
Pillow>=9.0.0
django-import-export>=3.0.0
openpyxl>=3.0.0
python-decouple>=3.8
reportlab>=4.0.0
gunicorn>=21.0.0
whitenoise>=6.5.0
psycopg2-binary>=2.9.0
dj-database-url>=2.1.0 